Refactor: Migrate regime data splitting to BaseStep pattern
