+++ src/training/steps/market_analysis/__init__.py
@@ -16,7 +16,7 @@ from .hdbscan_clustering import HDBSCANRegimeDiscoveryStep
 # Import migrated steps
 from .sr_detection import SRDetectionStep
 from .model_persistence_components.model_persistence_step import ModelPersistenceStep
-from .regime_data_splitting.regime_data_splitting_step import RegimeDataSplittingStep
+from .regime_data_splitting.regime_data_splitting_main import RegimeDataSplittingStep
 from .components.regime_models_training import RegimeModelsTrainingStep
 from .components.regime_ensemble_training import RegimeEnsembleTrainingStep
 from .components.sr_parameter_optimization import SRParameterOptimizationStep
